+++ src/main/java/com/carterz30cal/areas/Areas.java
@@ -1,0 +1,18 @@
+package com.carterz30cal.areas;
+
+import com.carterz30cal.areas.areas.GameAreaWaterway;
+
+public enum Areas {
+    WATERWAY(new GameAreaWaterway());
+    private final AbstractGameArea area;
+
+    Areas(AbstractGameArea area) {
+        this.area = area;
+        this.area.parent = this;
+    }
+
+    public AbstractGameArea getArea() {
+        return area;
+    }
+
+}
