+++ src/main/java/com/carterz30cal/areas/bosses/AbstractAreaBoss.java
@@ -1,0 +1,47 @@
+package com.carterz30cal.areas.bosses;
+
+import com.carterz30cal.areas.PlayerTeleport;
+import com.carterz30cal.areas.spawners.AbstractEnemySpawner;
+import com.carterz30cal.entities.player.GamePlayer;
+import org.bukkit.Location;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class AbstractAreaBoss extends AbstractEnemySpawner {
+    protected boolean active = false;
+
+    public void OnRightClick(GamePlayer player, Location clicked) {
+
+    }
+
+    public void OnLeftClick(GamePlayer player, Location clicked) {
+
+    }
+
+    public void OnPlayerDeath(GamePlayer player) {
+
+    }
+
+    public void OnTeleport(GamePlayer player, PlayerTeleport teleport) {
+
+    }
+
+    public boolean IsActive() {
+        return active;
+    }
+
+    public boolean IsPlayerInvolved(GamePlayer player) {
+        return false;
+    }
+
+    public String GetName(GamePlayer player) {
+        return "Boss";
+    }
+
+    public List<String> GetScoreboard(GamePlayer player) {
+        List<String> list = new ArrayList<String>();
+        list.add(GetName(player));
+        return list;
+    }
+}
