+++ src/main/java/com/carterz30cal/areas/quests/requirements/QuestRequirement.java
@@ -1,0 +1,7 @@
+package com.carterz30cal.areas.quests.requirements;
+
+import com.carterz30cal.entities.player.GamePlayer;
+
+public interface QuestRequirement {
+    boolean HasMetRequirements(GamePlayer player);
+}
