--- src/main/java/com/carterz30cal/areas2/Areas.java
@@ -1,18 +1,0 @@
-package com.carterz30cal.areas2;
-
-import com.carterz30cal.areas2.areas.GameAreaWaterway;
-
-public enum Areas {
-    WATERWAY(new GameAreaWaterway());
-    private final AbstractGameArea area;
-
-    Areas(AbstractGameArea area) {
-        this.area = area;
-        this.area.parent = this;
-    }
-
-    public AbstractGameArea getArea() {
-        return area;
-    }
-
-}
