--- src/main/java/com/carterz30cal/areas2/events/AbstractGameAreaEvent.java
@@ -1,9 +1,0 @@
-package com.carterz30cal.areas2.events;
-
-import com.carterz30cal.areas2.Areas;
-
-public abstract class AbstractGameAreaEvent {
-    public boolean isPossibleInArea(Areas area) {
-        return false;
-    }
-}
