--- src/main/java/com/carterz30cal/areas2/quests/requirements/QuestRequirement.java
@@ -1,7 +1,0 @@
-package com.carterz30cal.areas2.quests.requirements;
-
-import com.carterz30cal.entities.player.GamePlayer;
-
-public interface QuestRequirement {
-    boolean HasMetRequirements(GamePlayer player);
-}
