+++ src/main/java/com/carterz30cal/entities/interactable/GameQuestgiver.java
@@ -1,8 +1,8 @@
 package com.carterz30cal.entities.interactable;
 
-import com.carterz30cal.areas2.quests.Questgivers;
-import com.carterz30cal.areas2.quests.Quests;
-import com.carterz30cal.areas2.quests.rewards.QuestReward;
+import com.carterz30cal.areas.quests.Questgivers;
+import com.carterz30cal.areas.quests.Quests;
+import com.carterz30cal.areas.quests.rewards.QuestReward;
 import com.carterz30cal.entities.player.GamePlayer;
 import com.carterz30cal.main.Dungeons;
 import org.bukkit.Sound;
