+++ src/main/java/com/carterz30cal/events/GameEventHandler.java
@@ -1,6 +1,6 @@
 package com.carterz30cal.events;
 
-import com.carterz30cal.areas2.Areas;
+import com.carterz30cal.areas.Areas;
 import com.carterz30cal.entities.GameEnemy;
 import com.carterz30cal.entities.player.GamePlayer;
 import org.bukkit.Location;
