+++ src/main/java/com/carterz30cal/gui/QuestGUI.java
@@ -1,6 +1,6 @@
 package com.carterz30cal.gui;
 
-import com.carterz30cal.areas2.quests.Quests;
+import com.carterz30cal.areas.quests.Quests;
 import com.carterz30cal.entities.player.GamePlayer;
 import com.carterz30cal.items.ItemFactory;
 import org.bukkit.inventory.ItemStack;
